+++ core/src/test/scala/com/sandinh/couchbase/CB.scala
@@ -4,6 +4,6 @@ import javax.inject._
 
 @Singleton
 class CB @Inject() (val cluster: CBCluster) {
-  lazy val bk1 = cluster.openBucket("bk1")
-  lazy val bk2 = cluster.openBucket("acc")
+  lazy val bk1 = cluster.openBucketSync("bk1")
+  lazy val bk2 = cluster.openBucketSync("acc")
 }